Update Headlines and Images
The image elements were enclosed inside a div element in order to align them center.
The images were then sized down to be able to see them without zooming out.
The h1 headlines were also aligned to the center and made underlined,bold, and font increased to emphasize that it is the main headline.
The background of the website was also changed to a light brown shade to make the website look nicer than plain white.