Add CSS to HTML files
Center aligned h1 elements, put images and lists
into div elements and center aligned them as well.
Changed the background color so the page looked nicer
and more aestheic. Changed the size of the images to
have a width of 500px so the image wasn't so big
compared to the text. Made the return to homepage
link a button and changed the background color of it
so it matched with the background on the page.